RM12488: Fix multiple elements like view and translation

+++ axelor-business-project/src/main/java/com/axelor/apps/businessproject/service/projectgenerator/factory/ProjectGeneratorFactoryTaskTemplate.java
@@ -77,7 +77,7 @@ public class ProjectGeneratorFactoryTaskTemplate implements ProjectGeneratorFact
                 project,
                 root,
                 startDate,
-                orderLine.getQty().intValue());
+                orderLine.getQty());
         tasks.addAll(convertedTasks);
       }
     }
